+++ tests/test_doh.py
@@ -31,10 +31,6 @@ import dns.query
 import dns.rdatatype
 import dns.resolver
 
-if dns.query._have_requests:
-    import requests
-    from requests.exceptions import SSLError
-
 if dns.query._have_httpx:
     import httpx
 
@@ -42,18 +38,26 @@ import tests.util
 
 resolver_v4_addresses = []
 resolver_v6_addresses = []
+family = socket.AF_UNSPEC
 if tests.util.have_ipv4():
     resolver_v4_addresses = [
         "1.1.1.1",
         "8.8.8.8",
         # '9.9.9.9',
     ]
+    family = socket.AF_INET
 if tests.util.have_ipv6():
     resolver_v6_addresses = [
         "2606:4700:4700::1111",
         "2001:4860:4860::8888",
         # '2620:fe::fe',
     ]
+    if family == socket.AF_INET:
+        # we have both working, go back to UNSPEC
+        family = socket.AF_UNSPEC
+    else:
+        # v6 only
+        family = socket.AF_INET6
 
 KNOWN_ANYCAST_DOH_RESOLVER_URLS = [
     "https://cloudflare-dns.com/dns-query",
@@ -65,86 +69,6 @@ KNOWN_PAD_AWARE_DOH_RESOLVER_URLS = [
     "https://cloudflare-dns.com/dns-query",
     "https://dns.google/dns-query",
 ]
-
-
-@unittest.skipUnless(
-    dns.query._have_requests and tests.util.is_internet_reachable(),
-    "Python requests cannot be imported; no DNS over HTTPS (DOH)",
-)
-class DNSOverHTTPSTestCaseRequests(unittest.TestCase):
-    def setUp(self):
-        self.session = requests.sessions.Session()
-
-    def tearDown(self):
-        self.session.close()
-
-    def test_get_request(self):
-        nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
-        q = dns.message.make_query("example.com.", dns.rdatatype.A)
-        r = dns.query.https(
-            q, nameserver_url, session=self.session, post=False, timeout=4
-        )
-        self.assertTrue(q.is_response(r))
-
-    def test_post_request(self):
-        nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
-        q = dns.message.make_query("example.com.", dns.rdatatype.A)
-        r = dns.query.https(
-            q, nameserver_url, session=self.session, post=True, timeout=4
-        )
-        self.assertTrue(q.is_response(r))
-
-    def test_build_url_from_ip(self):
-        self.assertTrue(resolver_v4_addresses or resolver_v6_addresses)
-        if resolver_v4_addresses:
-            nameserver_ip = random.choice(resolver_v4_addresses)
-            q = dns.message.make_query("example.com.", dns.rdatatype.A)
-            # For some reason Google's DNS over HTTPS fails when you POST to
-            # https://8.8.8.8/dns-query
-            # So we're just going to do GET requests here
-            r = dns.query.https(
-                q, nameserver_ip, session=self.session, post=False, timeout=4
-            )
-
-            self.assertTrue(q.is_response(r))
-        if resolver_v6_addresses:
-            nameserver_ip = random.choice(resolver_v6_addresses)
-            q = dns.message.make_query("example.com.", dns.rdatatype.A)
-            r = dns.query.https(
-                q, nameserver_ip, session=self.session, post=False, timeout=4
-            )
-            self.assertTrue(q.is_response(r))
-
-    def test_bootstrap_address(self):
-        # We test this to see if v4 is available
-        if resolver_v4_addresses:
-            ip = "185.228.168.168"
-            invalid_tls_url = "https://{}/doh/family-filter/".format(ip)
-            valid_tls_url = "https://doh.cleanbrowsing.org/doh/family-filter/"
-            q = dns.message.make_query("example.com.", dns.rdatatype.A)
-            # make sure CleanBrowsing's IP address will fail TLS certificate
-            # check
-            with self.assertRaises(SSLError):
-                dns.query.https(q, invalid_tls_url, session=self.session, timeout=4)
-            # use host header
-            r = dns.query.https(
-                q, valid_tls_url, session=self.session, bootstrap_address=ip, timeout=4
-            )
-            self.assertTrue(q.is_response(r))
-
-    def test_new_session(self):
-        nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
-        q = dns.message.make_query("example.com.", dns.rdatatype.A)
-        r = dns.query.https(q, nameserver_url, timeout=4)
-        self.assertTrue(q.is_response(r))
-
-    def test_resolver(self):
-        res = dns.resolver.Resolver(configure=False)
-        res.nameservers = ["https://dns.google/dns-query"]
-        answer = res.resolve("dns.google", "A")
-        seen = set([rdata.address for rdata in answer])
-        self.assertTrue("8.8.8.8" in seen)
-        self.assertTrue("8.8.4.4" in seen)
 
 
 @unittest.skipUnless(
@@ -162,7 +86,12 @@ class DNSOverHTTPSTestCaseHttpx(unittest.TestCase):
         nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
         q = dns.message.make_query("example.com.", dns.rdatatype.A)
         r = dns.query.https(
-            q, nameserver_url, session=self.session, post=False, timeout=4
+            q,
+            nameserver_url,
+            session=self.session,
+            post=False,
+            timeout=4,
+            family=family,
         )
         self.assertTrue(q.is_response(r))
 
@@ -173,7 +102,12 @@ class DNSOverHTTPSTestCaseHttpx(unittest.TestCase):
             nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
             q = dns.message.make_query("example.com.", dns.rdatatype.A)
             r = dns.query.https(
-                q, nameserver_url, session=self.session, post=False, timeout=4
+                q,
+                nameserver_url,
+                session=self.session,
+                post=False,
+                timeout=4,
+                family=family,
             )
             self.assertTrue(q.is_response(r))
         finally:
@@ -183,7 +117,12 @@ class DNSOverHTTPSTestCaseHttpx(unittest.TestCase):
         nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)
         q = dns.message.make_query("example.com.", dns.rdatatype.A)
         r = dns.query.https(
-            q, nameserver_url, session=self.session, post=True, timeout=4
+            q,
+            nameserver_url,
+            session=self.session,
+            post=True,
+            timeout=4,
+            family=family,
         )
         self.assertTrue(q.is_response(r))
 
@@ -219,17 +158,15 @@ class DNSOverHTTPSTestCaseHttpx(unittest.TestCase):
             # check.
             with self.assertRaises(httpx.ConnectError):
                 dns.query.https(q, invalid_tls_url, session=self.session, timeout=4)
-            # We can't do the Host header and SNI magic with httpx, but
-            # we are demanding httpx be used by providing a session, so
-            # we should get a NoDOH exception.
-            with self.assertRaises(dns.query.NoDOH):
-                dns.query.https(
-                    q,
-                    valid_tls_url,
-                    session=self.session,
-                    bootstrap_address=ip,
-                    timeout=4,
-                )
+            # And if we don't mangle the URL, it should work.
+            r = dns.query.https(
+                q,
+                valid_tls_url,
+                session=self.session,
+                bootstrap_address=ip,
+                timeout=4,
+            )
+            self.assertTrue(q.is_response(r))
 
     def test_new_session(self):
         nameserver_url = random.choice(KNOWN_ANYCAST_DOH_RESOLVER_URLS)